week3/requirements.txt
aiohttp>=3.9.0
base58>=2.1.1
httpx>=0.27.0
pandas>=2.2.0
pyarrow>=15.0.0
duckdb>=0.10.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.32.0
matplotlib>=3.8.0
statsmodels>=0.14.0
